check min/max_zoom for tms using TileMapÃ metadata
