+++ requirements.txt
@@ -8,3 +8,5 @@ mercantile==1.1.6
 owslib==0.23.0
 validators==0.18.2
 regex==2021.3.17
+beautifulsoup4==4.9.3
+lxml==4.6.3
